requirements.txt
Flask
Pillow
cryptography
qrcode[pil]
